Log-Ingestion-and-Metrics/app/middleware/request_size_limit.py
from starlette.middleware.base import BaseHTTPMiddleware
from starlette.requests import Request
from starlette.responses import JSONResponse


class RequestSizeLimitMiddleware(BaseHTTPMiddleware):
    # 64 KB — suficiente para logs HAProxy (SDD §9.7.2)
    MAX_BODY_SIZE = 64 * 1024

    async def dispatch(self, request: Request, call_next):
        content_length = request.headers.get("content-length")
        if content_length and int(content_length) > self.MAX_BODY_SIZE:
            return JSONResponse(status_code=413, content={"error": "Request body too large"})
        return await call_next(request)
