Log-Ingestion-and-Metrics/requirements.txt
# Web framework
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9

# Data validation
pydantic==2.7.1
pydantic-settings==2.2.1

# Redis
redis[hiredis]==5.0.4

# Observability
opentelemetry-sdk==1.24.0
opentelemetry-instrumentation-fastapi==0.45b0
opentelemetry-instrumentation-redis==0.45b0
opentelemetry-exporter-otlp-proto-grpc==1.24.0
prometheus-fastapi-instrumentator==7.0.0
prometheus-client==0.20.0

# Testing
pytest==8.2.0
pytest-asyncio==0.23.7
pytest-cov==5.0.0
httpx==0.27.0
fakeredis[aioredis]==2.23.2
